Inject log export dependencies
